backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "apispec>=6.8.2",
    "apispec-webframeworks>=1.2.0",
    "apscheduler>=3.11.0",
    "click>=8.2.1",
    "cryptography>=45.0.3",
    "deprecated>=1.2.18",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "flask-migrate>=4.1.0",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-swagger-ui>=5.21.0",
    "gunicorn>=23.0.0",
    "marshmallow>=4.0.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-jaeger-thrift>=1.21.0",
    "opentelemetry-instrumentation-flask>=0.57b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "pymysql>=1.1.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-flask>=1.3.0",
    "pytest-mock>=3.14.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "werkzeug>=3.1.3",
    "pyyaml>=6.0.2",
]
